+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Validate if referenced `device` exists in provider configuration
 - Add `delete_mode` attribute to resources
 - Add `cache_rpf_oif` attribute to `iosxe_pim_vrf` resource and data source
+- Add `iosxe_ntp` resource and data source
 
 ## 0.1.17